+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   * Add requirement ruby >= 2.1
   * Drop support for Rails < 4.2
   * Replace `thread_safe` with `concurrent_ruby`
+  * Remove custom responder
 
 ## 0.4.3
   * Fix custom responder compatibility with responders 2.1 (itkin)